// extensions/xcode/src/models/simulator/xcode-simulator.model.ts
/**
 * A Xcode Simulator
 */
export interface XcodeSimulator {
  /**
   * The udid
   * Note: No typo! udid is the field which the simctl responds with
   */
  udid: string;
  /**
   * The name
   */
  name: string;
  /**
   * The runtime e.g. iOS-15-0, tvOS-15-0
   */
  runtime: string;
  /**
   * The state e.g. Shutdown, Booted
   */
  state: string;
  /**
   * The data file path
   */
  dataPath: string;
}